Ajustando classe

+++ Repositories/ReceitaRepository.cs
@@ -14,11 +14,11 @@ namespace receitas.Repositories
     public class ReceitaRepository : IReceitaRepository
     {
         public string ConnectionString ="Server=localhost;Port=3306;Database=receitas;Uid=jefferson;Pwd=jefferson23";
-        public IEnumerable<Receita> FindAll()
+        public IEnumerable<Receita> FindAllTrue()
         {
             using (IDbConnection dbConnection = new MySqlConnection(ConnectionString))
             {
-                string sQuery = " SELECT id, nome, ingredientes, modo_de_preparar, status FROM receita" ;
+                string sQuery = " SELECT id, nome, ingredientes, modo_de_preparar as ModoPreparar, status FROM receita WHERE status = true" ;
               
                 dbConnection.Open();
                 return dbConnection.Query<Receita>(sQuery);
@@ -30,12 +30,15 @@ namespace receitas.Repositories
             using (IDbConnection dbConnection = new MySqlConnection(ConnectionString))
             {
                 var receitaD = findById(ParametroId);
+                
+                if(receitaD != null)
+                {
+                     string sQuery = "DELETE FROM receita WHERE id = @id";
+                    dbConnection.Open();
 
-                string sQuery = "DELETE FROM receita WHERE id = @id";
-                dbConnection.Open();
+                    dbConnection.Execute(sQuery, new {id = ParametroId});  
 
-                dbConnection.Execute(sQuery, new {id = ParametroId});   
-
+                }
                 return receitaD;
             }
         }
@@ -43,7 +46,7 @@ namespace receitas.Repositories
         {
             using (IDbConnection dbConnection = new MySqlConnection(ConnectionString))
             {
-                string sQuery = "SELECT id, nome, ingredientes, modo_de_preparar , status FROM receita WHERE id = @id"; 
+                string sQuery = "SELECT id, nome, ingredientes, modo_de_preparar as ModoPreparar , status FROM receita WHERE id = @id"; 
 
                 try
                 {
@@ -61,17 +64,17 @@ namespace receitas.Repositories
             using (IDbConnection dbConnection = new MySqlConnection(ConnectionString))
             {       
                 string sQueryS = "SELECT LAST_INSERT_ID()";
-                string sQuery = "INSERT INTO receita(nome, ingredientes, modo_de_preparar,status) VALUES(@nome, @ingredientes, @modo_de_preparar, @status)";
+                string sQuery = "INSERT INTO receita(nome, ingredientes, modo_de_preparar,status) VALUES(@Nome, @Ingredientes, @ModoPreparar, @Status)";
                 
-                receita.status = true;
+                receita.Status = true;
 
                 dbConnection.Open();
 
                 dbConnection.Execute(sQuery, receita);   
 
-                receita.id = dbConnection.QueryFirst<int>(sQueryS);   
+                //receita.id = dbConnection.QueryFirst<int>(sQueryS);   
 
-                return receita;
+                return findById(dbConnection.QueryFirst<int>(sQueryS));
                
 
             }
@@ -85,7 +88,7 @@ namespace receitas.Repositories
             string sQuery = "UPDATE receita SET status = @status WHERE id = @id";   
             
             
-            var id = receita.id;
+            var id = receita.Id;
             var existe = findById(id);   
 
             if(existe != null)
@@ -96,7 +99,7 @@ namespace receitas.Repositories
                 
             }          
 
-            return existe = findById(id);
+            return existe;
             }
 
         }
